[Server, Common, Client] Made networking codes integers to stop with annoying casst (Fixed #38)

+++ client-source/World/World.cpp
@@ -148,7 +148,7 @@ void World::requestChunk(sf::Vector2i pos)
 {
     sf::Packet request;
 
-    request << (unsigned short)Networking::CtoS::RequestChunk;
+    request << Networking::CtoS::RequestChunk;
     request << pos.x << pos.y;
 
     state_game.sendToServer(request);
