+++ common-source/Items/GameItems/BallItem.cpp
@@ -25,8 +25,8 @@ void BallItem::use(ItemStack& stack, World& world, sf::Vector2i click_pos, Playe
     #ifndef CLIENT_SIDE
     sf::Packet set;
 
-    set << (unsigned short)Networking::StoC::InventoryUpdate;
-    set << (unsigned short)InventoryUpdates::StoC::AddStack;
+    set << Networking::StoC::InventoryUpdate;
+    set << InventoryUpdates::StoC::AddStack;
     set << 0;   //We set the first slot of the inventory (hand)
     set << stack.getInt();
